* Kā ē-té sî-kan-pió suá--khí-lâi
