+++ src/SiKanPio/SiKanHangBok.js
@@ -1,7 +1,5 @@
-import React from 'react'
-import './SiKanPio.css'
 
-const sikan=[
+const SiKanHangBok = [
   {
     jit: '2019.11.01 拜五',
     hangbok: [
@@ -176,23 +174,4 @@ const sikan=[
 
 ]
 
-const lueiong=sikan.map((kang,i) =>
-  <div key={i}>
-    <span className='jit'>{kang.jit}</span><br/>
-	{kang.hangbok.map(tiunn =>
-      <div key={tiunn.sikan}>
-        <span className='sikan'>{tiunn.sikan}</span><br/>
-        <span className='tebok'>{tiunn.tebok}</span><br/>
-        <span className='sootsai'>{tiunn.sootsai}</span>
-      </div>
-	)}
-  </div>
-)
-const SiKanPio = () => (
-  <div className='sikanpio ui text container'>
-    <h1 className="ui inverted center aligned header">時間表</h1>
-    {lueiong}
-  </div>
-)
-
-export default SiKanPio
+export default SiKanHangBok
